pyproject.toml
[project]
name = "telegram-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.10.1",
    "flask[async]>=3.1.2",
    "pydantic>=2.12.5",
    "pyrogram>=2.0.106",
    "python-dotenv>=1.2.1",
    "telethon>=1.42.0",
    "tgcrypto>=1.2.5",
]
